# packages/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: JeevaSetu Karnataka API
  version: 0.1.0
  description: >-
    A clinician-supervised, closed-loop maternal referral safety layer for
    high-risk pregnancies in Karnataka. Tracks referrals from creation through
    acknowledgement, acceptance, transport, arrival, clinical disposition,
    discharge, follow-up, and closure with GapSense classification.
servers:
  - url: /api/v1
    description: Current API v1 endpoint
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users (Admin only)
      operationId: listUsers
      parameters:
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Role'
        - name: facilityId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}:
    patch:
      tags:
        - Users
      summary: Update user details or active status (Admin only)
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                role:
                  $ref: '#/components/schemas/Role'
                facilityId:
                  type: string
                  format: uuid
                  nullable: true
                isActive:
                  type: boolean
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /blackspot/summary:
    get:
      tags:
        - Blackspot
      summary: Read-only district referral blackspot intelligence summary
      operationId: getBlackspotSummary
      parameters:
        - name: district
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Aggregated capacity and gap signals with pilot disclaimer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlackspotSummary'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique UUID token per mutation request to ensure safe replay
      schema:
        type: string
        format: uuid
  responses:
...
    ForbiddenError:
      description: User lacks sufficient role or geographic scope for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Requested entity was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Role:
      type: string
      enum:
        - FRONTLINE_WORKER
        - SENDING_FACILITY
        - RECEIVING_FACILITY
        - CLINICIAN
        - DISTRICT_SUPERVISOR
        - ADMINISTRATOR
        - CLINICAL_ADMINISTRATOR
...
    GapPhase:
      type: string
      enum:
        - ACKNOWLEDGEMENT
        - TRANSPORT
        - CAPACITY
        - DISPOSITION
        - FOLLOW_UP
    GapCauseClass:
      type: string
      enum:
        - CAPACITY
        - PROCESS
        - COMMUNICATION
        - UNDETERMINED
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        facilityId:
          type: string
          format: uuid
          nullable: true
        phone:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BlackspotSummary:
      type: object
      properties:
        district:
          type: string
        pilotDisclaimer:
          type: string
          default: 'Pilot-period, synthetic-data only. Aggregated output suppressed below minimum case threshold.'
        minimumThresholdMet:
          type: boolean
        blackspots:
          type: array
          items:
            type: object
            properties:
              facilityId:
                type: string
                format: uuid
              facilityName:
                type: string
              topFailurePhase:
                $ref: '#/components/schemas/GapPhase'
              topCauseClass:
                $ref: '#/components/schemas/GapCauseClass'
              rejectionRate:
                type: number
              totalReferrals:
                type: integer
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - requestId
        - timestamp
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time